# openapi.yaml
openapi: 3.0.3
info:
  title: Интернет-магазин API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /products:
    get:
      summary: Получить список товаров
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    price:
                      type: number
                    description:
                      type: string
                    categories:
                      type: array
                      items:
                        type: string
    post:
      summary: Добавить новый товар
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: string
                categories:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Товар добавлен
  /products/{id}:
    get:
      summary: Получить товар по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  price:
                    type: number
                  description:
                    type: string
                  categories:
                    type: array
                    items:
                      type: string
        '404':
          description: Товар не найден
    put:
      summary: Обновить товар
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: string
                categories:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Товар обновлён
        '404':
          description: Товар не найден
    delete:
      summary: Удалить товар
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Товар удалён
        '404':
          description: Товар не найден